Crate rusoto_worklink

Amazon WorkLink is a cloud-based service that provides secure access to internal websites and web apps from iOS phones. In a single step, your users, such as employees, can access internal websites as efficiently as they access any other public website. They enter a URL in their web browser, or choose a link to an internal website in an email. Amazon WorkLink authenticates the user's access and securely renders authorized internal web content in a secure rendering service in the AWS cloud. Amazon WorkLink doesn't download or store any internal web content on mobile devices.

If you're using the service, you're probably looking for WorklinkClient and Worklink.

Structs

AssociateDomainRequest

AssociateDomainResponse

AssociateWebsiteAuthorizationProviderRequest

AssociateWebsiteAuthorizationProviderResponse

AssociateWebsiteCertificateAuthorityRequest

AssociateWebsiteCertificateAuthorityResponse

CreateFleetRequest

CreateFleetResponse

DeleteFleetRequest

DeleteFleetResponse

DescribeAuditStreamConfigurationRequest

DescribeAuditStreamConfigurationResponse

DescribeCompanyNetworkConfigurationRequest

DescribeCompanyNetworkConfigurationResponse

DescribeDevicePolicyConfigurationRequest

DescribeDevicePolicyConfigurationResponse

DescribeDeviceRequest

DescribeDeviceResponse

DescribeDomainRequest

DescribeDomainResponse

DescribeFleetMetadataRequest

DescribeFleetMetadataResponse

DescribeIdentityProviderConfigurationRequest

DescribeIdentityProviderConfigurationResponse

DescribeWebsiteCertificateAuthorityRequest

DescribeWebsiteCertificateAuthorityResponse

DeviceSummary
The summary of devices.

DisassociateDomainRequest

DisassociateDomainResponse

DisassociateWebsiteAuthorizationProviderRequest

DisassociateWebsiteAuthorizationProviderResponse

DisassociateWebsiteCertificateAuthorityRequest

DisassociateWebsiteCertificateAuthorityResponse

DomainSummary
The summary of the domain.

FleetSummary
The summary of the fleet.

ListDevicesRequest

ListDevicesResponse

ListDomainsRequest

ListDomainsResponse

ListFleetsRequest

ListFleetsResponse

ListWebsiteAuthorizationProvidersRequest

ListWebsiteAuthorizationProvidersResponse

ListWebsiteCertificateAuthoritiesRequest

ListWebsiteCertificateAuthoritiesResponse

RestoreDomainAccessRequest

RestoreDomainAccessResponse

RevokeDomainAccessRequest

RevokeDomainAccessResponse

SignOutUserRequest

SignOutUserResponse

UpdateAuditStreamConfigurationRequest

UpdateAuditStreamConfigurationResponse

UpdateCompanyNetworkConfigurationRequest

UpdateCompanyNetworkConfigurationResponse

UpdateDevicePolicyConfigurationRequest

UpdateDevicePolicyConfigurationResponse

UpdateDomainMetadataRequest

UpdateDomainMetadataResponse

UpdateFleetMetadataRequest

UpdateFleetMetadataResponse

UpdateIdentityProviderConfigurationRequest

UpdateIdentityProviderConfigurationResponse

WebsiteAuthorizationProviderSummary
The summary of the website authorization provider.

WebsiteCaSummary
The summary of the certificate authority (CA).

WorklinkClient
A client for the WorkLink API.

Enums

AssociateDomainError
Errors returned by AssociateDomain

AssociateWebsiteAuthorizationProviderError
Errors returned by AssociateWebsiteAuthorizationProvider

AssociateWebsiteCertificateAuthorityError
Errors returned by AssociateWebsiteCertificateAuthority

CreateFleetError
Errors returned by CreateFleet

DeleteFleetError
Errors returned by DeleteFleet

DescribeAuditStreamConfigurationError
Errors returned by DescribeAuditStreamConfiguration

DescribeCompanyNetworkConfigurationError
Errors returned by DescribeCompanyNetworkConfiguration

DescribeDeviceError
Errors returned by DescribeDevice

DescribeDevicePolicyConfigurationError
Errors returned by DescribeDevicePolicyConfiguration

DescribeDomainError
Errors returned by DescribeDomain

DescribeFleetMetadataError
Errors returned by DescribeFleetMetadata

DescribeIdentityProviderConfigurationError
Errors returned by DescribeIdentityProviderConfiguration

DescribeWebsiteCertificateAuthorityError
Errors returned by DescribeWebsiteCertificateAuthority

DisassociateDomainError
Errors returned by DisassociateDomain

DisassociateWebsiteAuthorizationProviderError
Errors returned by DisassociateWebsiteAuthorizationProvider

DisassociateWebsiteCertificateAuthorityError
Errors returned by DisassociateWebsiteCertificateAuthority

ListDevicesError
Errors returned by ListDevices

ListDomainsError
Errors returned by ListDomains

ListFleetsError
Errors returned by ListFleets

ListWebsiteAuthorizationProvidersError
Errors returned by ListWebsiteAuthorizationProviders

ListWebsiteCertificateAuthoritiesError
Errors returned by ListWebsiteCertificateAuthorities

RestoreDomainAccessError
Errors returned by RestoreDomainAccess

RevokeDomainAccessError
Errors returned by RevokeDomainAccess

SignOutUserError
Errors returned by SignOutUser

UpdateAuditStreamConfigurationError
Errors returned by UpdateAuditStreamConfiguration

UpdateCompanyNetworkConfigurationError
Errors returned by UpdateCompanyNetworkConfiguration

UpdateDevicePolicyConfigurationError
Errors returned by UpdateDevicePolicyConfiguration

UpdateDomainMetadataError
Errors returned by UpdateDomainMetadata

UpdateFleetMetadataError
Errors returned by UpdateFleetMetadata

UpdateIdentityProviderConfigurationError
Errors returned by UpdateIdentityProviderConfiguration

Traits

Worklink
Trait representing the capabilities of the WorkLink API. WorkLink clients implement this trait.