Struct rusoto_worklink::WorklinkClient

pub struct WorklinkClient { /* fields omitted */ }

A client for the WorkLink API.

Methods

impl WorklinkClient

pub fn new(region: Region) -> WorklinkClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> WorklinkClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,