Trait rusoto_worklink::Worklink

pub trait Worklink {
    fn associate_domain(
        &self, 
        input: AssociateDomainRequest
    ) -> RusotoFuture<AssociateDomainResponse, AssociateDomainError>;

    fn associate_website_authorization_provider(
        &self, 
        input: AssociateWebsiteAuthorizationProviderRequest
    ) -> RusotoFuture<AssociateWebsiteAuthorizationProviderResponse, AssociateWebsiteAuthorizationProviderError>;

    fn associate_website_certificate_authority(
        &self, 
        input: AssociateWebsiteCertificateAuthorityRequest
    ) -> RusotoFuture<AssociateWebsiteCertificateAuthorityResponse, AssociateWebsiteCertificateAuthorityError>;

    fn create_fleet(
        &self, 
        input: CreateFleetRequest
    ) -> RusotoFuture<CreateFleetResponse, CreateFleetError>;

    fn delete_fleet(
        &self, 
        input: DeleteFleetRequest
    ) -> RusotoFuture<DeleteFleetResponse, DeleteFleetError>;

    fn describe_audit_stream_configuration(
        &self, 
        input: DescribeAuditStreamConfigurationRequest
    ) -> RusotoFuture<DescribeAuditStreamConfigurationResponse, DescribeAuditStreamConfigurationError>;

    fn describe_company_network_configuration(
        &self, 
        input: DescribeCompanyNetworkConfigurationRequest
    ) -> RusotoFuture<DescribeCompanyNetworkConfigurationResponse, DescribeCompanyNetworkConfigurationError>;

    fn describe_device(
        &self, 
        input: DescribeDeviceRequest
    ) -> RusotoFuture<DescribeDeviceResponse, DescribeDeviceError>;

    fn describe_device_policy_configuration(
        &self, 
        input: DescribeDevicePolicyConfigurationRequest
    ) -> RusotoFuture<DescribeDevicePolicyConfigurationResponse, DescribeDevicePolicyConfigurationError>;

    fn describe_domain(
        &self, 
        input: DescribeDomainRequest
    ) -> RusotoFuture<DescribeDomainResponse, DescribeDomainError>;

    fn describe_fleet_metadata(
        &self, 
        input: DescribeFleetMetadataRequest
    ) -> RusotoFuture<DescribeFleetMetadataResponse, DescribeFleetMetadataError>;

    fn describe_identity_provider_configuration(
        &self, 
        input: DescribeIdentityProviderConfigurationRequest
    ) -> RusotoFuture<DescribeIdentityProviderConfigurationResponse, DescribeIdentityProviderConfigurationError>;

    fn describe_website_certificate_authority(
        &self, 
        input: DescribeWebsiteCertificateAuthorityRequest
    ) -> RusotoFuture<DescribeWebsiteCertificateAuthorityResponse, DescribeWebsiteCertificateAuthorityError>;

    fn disassociate_domain(
        &self, 
        input: DisassociateDomainRequest
    ) -> RusotoFuture<DisassociateDomainResponse, DisassociateDomainError>;

    fn disassociate_website_authorization_provider(
        &self, 
        input: DisassociateWebsiteAuthorizationProviderRequest
    ) -> RusotoFuture<DisassociateWebsiteAuthorizationProviderResponse, DisassociateWebsiteAuthorizationProviderError>;

    fn disassociate_website_certificate_authority(
        &self, 
        input: DisassociateWebsiteCertificateAuthorityRequest
    ) -> RusotoFuture<DisassociateWebsiteCertificateAuthorityResponse, DisassociateWebsiteCertificateAuthorityError>;

    fn list_devices(
        &self, 
        input: ListDevicesRequest
    ) -> RusotoFuture<ListDevicesResponse, ListDevicesError>;

    fn list_domains(
        &self, 
        input: ListDomainsRequest
    ) -> RusotoFuture<ListDomainsResponse, ListDomainsError>;

    fn list_fleets(
        &self, 
        input: ListFleetsRequest
    ) -> RusotoFuture<ListFleetsResponse, ListFleetsError>;

    fn list_website_authorization_providers(
        &self, 
        input: ListWebsiteAuthorizationProvidersRequest
    ) -> RusotoFuture<ListWebsiteAuthorizationProvidersResponse, ListWebsiteAuthorizationProvidersError>;

    fn list_website_certificate_authorities(
        &self, 
        input: ListWebsiteCertificateAuthoritiesRequest
    ) -> RusotoFuture<ListWebsiteCertificateAuthoritiesResponse, ListWebsiteCertificateAuthoritiesError>;

    fn restore_domain_access(
        &self, 
        input: RestoreDomainAccessRequest
    ) -> RusotoFuture<RestoreDomainAccessResponse, RestoreDomainAccessError>;

    fn revoke_domain_access(
        &self, 
        input: RevokeDomainAccessRequest
    ) -> RusotoFuture<RevokeDomainAccessResponse, RevokeDomainAccessError>;

    fn sign_out_user(
        &self, 
        input: SignOutUserRequest
    ) -> RusotoFuture<SignOutUserResponse, SignOutUserError>;

    fn update_audit_stream_configuration(
        &self, 
        input: UpdateAuditStreamConfigurationRequest
    ) -> RusotoFuture<UpdateAuditStreamConfigurationResponse, UpdateAuditStreamConfigurationError>;

    fn update_company_network_configuration(
        &self, 
        input: UpdateCompanyNetworkConfigurationRequest
    ) -> RusotoFuture<UpdateCompanyNetworkConfigurationResponse, UpdateCompanyNetworkConfigurationError>;

    fn update_device_policy_configuration(
        &self, 
        input: UpdateDevicePolicyConfigurationRequest
    ) -> RusotoFuture<UpdateDevicePolicyConfigurationResponse, UpdateDevicePolicyConfigurationError>;

    fn update_domain_metadata(
        &self, 
        input: UpdateDomainMetadataRequest
    ) -> RusotoFuture<UpdateDomainMetadataResponse, UpdateDomainMetadataError>;

    fn update_fleet_metadata(
        &self, 
        input: UpdateFleetMetadataRequest
    ) -> RusotoFuture<UpdateFleetMetadataResponse, UpdateFleetMetadataError>;

    fn update_identity_provider_configuration(
        &self, 
        input: UpdateIdentityProviderConfigurationRequest
    ) -> RusotoFuture<UpdateIdentityProviderConfigurationResponse, UpdateIdentityProviderConfigurationError>;
}

Trait representing the capabilities of the WorkLink API. WorkLink clients implement this trait.

Required methods

fn associate_domain(
&self,
input: AssociateDomainRequest
) -> RusotoFuture<AssociateDomainResponse, AssociateDomainError>

Specifies a domain to be associated to Amazon WorkLink.

fn associate_website_authorization_provider(
&self,
input: AssociateWebsiteAuthorizationProviderRequest
) -> RusotoFuture<AssociateWebsiteAuthorizationProviderResponse, AssociateWebsiteAuthorizationProviderError>

Associates a website authorization provider with a specified fleet. This is used to authorize users against associated websites in the company network.

fn associate_website_certificate_authority(
&self,
input: AssociateWebsiteCertificateAuthorityRequest
) -> RusotoFuture<AssociateWebsiteCertificateAuthorityResponse, AssociateWebsiteCertificateAuthorityError>

Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network.

fn create_fleet(
&self,
input: CreateFleetRequest
) -> RusotoFuture<CreateFleetResponse, CreateFleetError>

Creates a fleet. A fleet consists of resources and the configuration that delivers associated websites to authorized users who download and set up the Amazon WorkLink app.

fn delete_fleet(
&self,
input: DeleteFleetRequest
) -> RusotoFuture<DeleteFleetResponse, DeleteFleetError>

Deletes a fleet. Prevents users from accessing previously associated websites.

fn describe_audit_stream_configuration(
&self,
input: DescribeAuditStreamConfigurationRequest
) -> RusotoFuture<DescribeAuditStreamConfigurationResponse, DescribeAuditStreamConfigurationError>

Describes the configuration for delivering audit streams to the customer account.

fn describe_company_network_configuration(
&self,
input: DescribeCompanyNetworkConfigurationRequest
) -> RusotoFuture<DescribeCompanyNetworkConfigurationResponse, DescribeCompanyNetworkConfigurationError>

Describes the networking configuration to access the internal websites associated with the specified fleet.

fn describe_device(
&self,
input: DescribeDeviceRequest
) -> RusotoFuture<DescribeDeviceResponse, DescribeDeviceError>

Provides information about a user's device.

fn describe_device_policy_configuration(
&self,
input: DescribeDevicePolicyConfigurationRequest
) -> RusotoFuture<DescribeDevicePolicyConfigurationResponse, DescribeDevicePolicyConfigurationError>

Describes the device policy configuration for the specified fleet.

fn describe_domain(
&self,
input: DescribeDomainRequest
) -> RusotoFuture<DescribeDomainResponse, DescribeDomainError>

Provides information about the domain.

fn describe_fleet_metadata(
&self,
input: DescribeFleetMetadataRequest
) -> RusotoFuture<DescribeFleetMetadataResponse, DescribeFleetMetadataError>

Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details.

fn describe_identity_provider_configuration(
&self,
input: DescribeIdentityProviderConfigurationRequest
) -> RusotoFuture<DescribeIdentityProviderConfigurationResponse, DescribeIdentityProviderConfigurationError>

Describes the identity provider configuration of the specified fleet.

fn describe_website_certificate_authority(
&self,
input: DescribeWebsiteCertificateAuthorityRequest
) -> RusotoFuture<DescribeWebsiteCertificateAuthorityResponse, DescribeWebsiteCertificateAuthorityError>

Provides information about the certificate authority.

fn disassociate_domain(
&self,
input: DisassociateDomainRequest
) -> RusotoFuture<DisassociateDomainResponse, DisassociateDomainError>

Disassociates a domain from Amazon WorkLink. End users lose the ability to access the domain with Amazon WorkLink.

fn disassociate_website_authorization_provider(
&self,
input: DisassociateWebsiteAuthorizationProviderRequest
) -> RusotoFuture<DisassociateWebsiteAuthorizationProviderResponse, DisassociateWebsiteAuthorizationProviderError>

Disassociates a website authorization provider from a specified fleet. After the disassociation, users can't load any associated websites that require this authorization provider.

fn disassociate_website_certificate_authority(
&self,
input: DisassociateWebsiteCertificateAuthorityRequest
) -> RusotoFuture<DisassociateWebsiteCertificateAuthorityResponse, DisassociateWebsiteCertificateAuthorityError>

Removes a certificate authority (CA).

fn list_devices(
&self,
input: ListDevicesRequest
) -> RusotoFuture<ListDevicesResponse, ListDevicesError>

Retrieves a list of devices registered with the specified fleet.

fn list_domains(
&self,
input: ListDomainsRequest
) -> RusotoFuture<ListDomainsResponse, ListDomainsError>

Retrieves a list of domains associated to a specified fleet.

fn list_fleets(
&self,
input: ListFleetsRequest
) -> RusotoFuture<ListFleetsResponse, ListFleetsError>

Retrieves a list of fleets for the current account and Region.

fn list_website_authorization_providers(
&self,
input: ListWebsiteAuthorizationProvidersRequest
) -> RusotoFuture<ListWebsiteAuthorizationProvidersResponse, ListWebsiteAuthorizationProvidersError>

Retrieves a list of website authorization providers associated with a specified fleet.

fn list_website_certificate_authorities(
&self,
input: ListWebsiteCertificateAuthoritiesRequest
) -> RusotoFuture<ListWebsiteCertificateAuthoritiesResponse, ListWebsiteCertificateAuthoritiesError>

Retrieves a list of certificate authorities added for the current account and Region.

fn restore_domain_access(
&self,
input: RestoreDomainAccessRequest
) -> RusotoFuture<RestoreDomainAccessResponse, RestoreDomainAccessError>

Moves a domain to ACTIVE status if it was in the INACTIVE status.

fn revoke_domain_access(
&self,
input: RevokeDomainAccessRequest
) -> RusotoFuture<RevokeDomainAccessResponse, RevokeDomainAccessError>

Moves a domain to INACTIVE status if it was in the ACTIVE status.

fn sign_out_user(
&self,
input: SignOutUserRequest
) -> RusotoFuture<SignOutUserResponse, SignOutUserError>

Signs the user out from all of their devices. The user can sign in again if they have valid credentials.

fn update_audit_stream_configuration(
&self,
input: UpdateAuditStreamConfigurationRequest
) -> RusotoFuture<UpdateAuditStreamConfigurationResponse, UpdateAuditStreamConfigurationError>

Updates the audit stream configuration for the fleet.

fn update_company_network_configuration(
&self,
input: UpdateCompanyNetworkConfigurationRequest
) -> RusotoFuture<UpdateCompanyNetworkConfigurationResponse, UpdateCompanyNetworkConfigurationError>

Updates the company network configuration for the fleet.

fn update_device_policy_configuration(
&self,
input: UpdateDevicePolicyConfigurationRequest
) -> RusotoFuture<UpdateDevicePolicyConfigurationResponse, UpdateDevicePolicyConfigurationError>

Updates the device policy configuration for the fleet.

fn update_domain_metadata(
&self,
input: UpdateDomainMetadataRequest
) -> RusotoFuture<UpdateDomainMetadataResponse, UpdateDomainMetadataError>

Updates domain metadata, such as DisplayName.

fn update_fleet_metadata(
&self,
input: UpdateFleetMetadataRequest
) -> RusotoFuture<UpdateFleetMetadataResponse, UpdateFleetMetadataError>

Updates fleet metadata, such as DisplayName.

fn update_identity_provider_configuration(
&self,
input: UpdateIdentityProviderConfigurationRequest
) -> RusotoFuture<UpdateIdentityProviderConfigurationResponse, UpdateIdentityProviderConfigurationError>

Updates the identity provider configuration for the fleet.

Loading content...